+++ aeolus/util/__init__.py
@@ -1,5 +1,0 @@
-"""Submodule with various utility functions used elsewhere in aeolus."""
-from .text import fmt_lonlat, subplot_label_generator
-
-
-__all__ = ("fmt_lonlat", "subplot_label_generator")