+++ conda.recipe/conda_upload.sh
@@ -5,8 +5,8 @@ PYCODE="import ${PKG_NAME}; print(${PKG_NAME}.__version__)"
 export VERSION=`python -c "${PYCODE}"`
 
 # OS=$TRAVIS_OS_NAME-64
-export CONDA_BLD_PATH=~/conda-bld
-mkdir $CONDA_BLD_PATH
+export CONDA_BLD_PATH=$HOME/conda-bld
+mkdir -p $CONDA_BLD_PATH
 conda config --set anaconda_upload no
 conda config --add channels conda-forge
 conda build --no-test .